runtime: patching testing/testing.go for go1.24
